requirements.txt
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-WTF==0.14.2
greenstalk==1.0.1
idna==2.8
importlib-metadata==1.2.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==8.0.2
pkg-resources==0.0.0
psycopg2==2.8.4
pyrsistent==0.15.6
python-dotenv==0.10.3
requests==2.22.0
selenium==3.141.0
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
WTForms==2.2.1
zipp==0.6.0
